create delete query
