+++ package.json
@@ -38,9 +38,11 @@
     "supertest-as-promised": "^4.0.2"
   },
   "dependencies": {
+    "cron": "^1.6.0",
     "express": "^4.16.3",
     "express-winston": "^2.6.0",
     "knex": "^0.15.2",
+    "moment": "^2.22.2",
     "pg": "^6.4.2",
     "sinon": "^2.4.1",
     "winston": "^2.4.4"
